+++ packages/seedless-onboarding-controller/src/SeedlessOnboardingController.ts
@@ -329,9 +329,6 @@ export class SeedlessOnboardingController<EncryptionKey> extends BaseController<
           rawToprfPwEncryptionKey: pwEncKey,
           rawToprfAuthKeyPair: authKeyPair,
         });
-        this.#persistAuthPubKey({
-          authPubKey: authKeyPair.pk,
-        });
       };
 
       await this.#executeWithTokenRefresh(
@@ -429,10 +426,6 @@ export class SeedlessOnboardingController<EncryptionKey> extends BaseController<
             rawToprfPwEncryptionKey: pwEncKey,
             rawToprfAuthKeyPair: authKeyPair,
           });
-
-          this.#persistAuthPubKey({
-            authPubKey: authKeyPair.pk,
-          });
         }
 
         return secrets;
@@ -491,9 +484,6 @@ export class SeedlessOnboardingController<EncryptionKey> extends BaseController<
           rawToprfAuthKeyPair: newAuthKeyPair,
         });
 
-        this.#persistAuthPubKey({
-          authPubKey: newAuthKeyPair.pk,
-        });
         this.#resetPasswordOutdatedCache();
 
         // store the keyring encryption key if it exists
@@ -640,10 +630,7 @@ export class SeedlessOnboardingController<EncryptionKey> extends BaseController<
           rawToprfPwEncryptionKey: pwEncKey,
           rawToprfAuthKeyPair: authKeyPair,
         });
-        // persist the latest global password authPubKey
-        this.#persistAuthPubKey({
-          authPubKey: authKeyPair.pk,
-        });
+
         this.#resetPasswordOutdatedCache();
       };
       return await this.#executeWithTokenRefresh(
@@ -1355,6 +1342,11 @@ export class SeedlessOnboardingController<EncryptionKey> extends BaseController<
       password,
       serializedVaultData,
       pwEncKey: rawToprfPwEncryptionKey,
+    });
+
+    // update the authPubKey in the state
+    this.#persistAuthPubKey({
+      authPubKey: rawToprfAuthKeyPair.pk,
     });
   }